requirements.txt
numpy
pyyaml
pytorch
torchvision
scikit-image
opencv-python
Cython